Fixed type, fixes #1
Fool of a Took. I struggled wit hthis so hard, trying to figure out why it wasn't working for some of the test cases. For the life of me I couldn't figure it out. Then the other day I thought.... I'll just double check that I've put the number of days for each month in right. Double checked it and..... I've put a 1 instead of 11 which meant that my code thought January had 30 days and November had 31. DOH.... Now it works perfectly of course. *KICKS SELF IN HEAD*